+++ db/migrate/20120830045204_create_definitions.rb
@@ -1,0 +1,10 @@
+class CreateDefinitions < ActiveRecord::Migration
+  def change
+    create_table :definitions do |t|
+      t.string :name
+      t.text :content
+
+      t.timestamps
+    end
+  end
+end