+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.1.0
 
 - definitions are now stored in the database.
+- No more `Mastermind.define {...}` block needed for definitions.
+- audit trail of definitions.